rename code.rs to expr.rs
reflects its more limited role
